+++ actfw_core/_private/schema/agent_app_protocol.py
@@ -101,7 +101,7 @@ class CommandResponse:
 class ServiceKind(enum.Enum):
     RS_256 = 0
     STOP_ACT = 1
-    DEVICE_SHUTDOWN = 2
+    SHUTDOWN_DEVICE = 2
 
 
 @dataclass(frozen=True, eq=False)